Begun tidying up output, and renaming stages
Could probably do with moving source into separate folder
Combining internal structure with keyboard top causes non-manifold
mesh issues in OpenSCAD, so it's done in meshlab instead. It probably
generates the same issues, but it doesn't complan, and it still seems
to slice anyway, so for now I'm pretending it doesn't exist.